requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
pandas
openpyxl
python-dotenv
